pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/factorialhr/__init__.py"

[project]
name = "factorialhr"
dynamic = ["version"]
authors = [
    { name = "Leon Budnick", email = "y6q6ea9w@mail-proxy.org" },
]
description = "Python package for the api of FactorialHR"
readme = "README.md"
requires-python = ">=3.11,<3.14"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
keywords = [
    "FactorialHR",
    "HR",
]
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
]

[dependency-groups]
dev = [
    "ruff>=0.9.6",
    "pyright[nodejs]>=1.1.391",
]

[project.optional-dependencies]
cli = [
    "cloup>=3.0.6",
    'aiohttp>=3.11.12',
    'python-dateutil>=2.9.0',
    'tabulate>=0.9.0',
]

[project.scripts]
factorialhr = 'factorialhr._cli:entrypoint'

[project.urls]
"Repository" = "https://github.com/leon1995/factorialhr"
"Bug Tracker" = "https://github.com/leon1995/factorialhr/issues"

[tool.hatch.build.targets.sdist]
packages = ["src/factorialhr"]

[tool.ruff]
lint.select = ["ALL"]
lint.extend-ignore = [
    "ANN002",
    "ANN003",
    "ANN401",
    "ANN401",
    "C408",
    "C409",
    "D100",
    "D105",
    "D107",
    "D203",
    "D213",
    "EM101",
    "FIX002",
    "S101",
    "S311",
    "SIM102",
    "T201",
    "TD002",
    "TD003",
    "TRY003",
]

line-length = 120  # https://docs.astral.sh/ruff/settings/#line-length
src = ["src"]  # https://docs.astral.sh/ruff/settings/#src

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true # https://docs.astral.sh/ruff/settings/#allow-star-arg-any
suppress-none-returning = true # https://docs.astral.sh/ruff/settings/#suppress-none-returning

[tool.ruff.lint.flake8-comprehensions]
allow-dict-calls-with-keyword-arguments = true  # https://docs.astral.sh/ruff/settings/#allow-dict-calls-with-keyword-arguments

[tool.ruff.lint.pycodestyle]
max-doc-length = 120 # https://docs.astral.sh/ruff/settings/#max-doc-length

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["D104"]
"**/endpoints/*.py" = ["D101"]

[tool.pyright]
ignore = ["examples/"]
